requirements.txt
asgiref==3.6.0
dj-database-url==2.1.0
Django==4.1.5
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.9.0
whitenoise==6.6.0
